timlin fix bugs

+++ app/Http/Controllers/GameController.php
@@ -35,9 +35,6 @@ class GameController extends Controller
     public function start(Request $request){
         return view('start');
     }
-    public function finish(){
-        return view('finish');
-    }
 
     public function run(Request $request){
         //基本資料
@@ -474,18 +471,16 @@ class GameController extends Controller
             };
         }
         $game_processes = game_process::where('user_id',$user_id)->get();
-        return view('tim的',[
+        return view('monthlyevent',[
             'game_processes' => $game_processes,
             'accomplish_achievements' => $accomplish_achievements,
         ]);
     }
-    public function end_game(Request $request){
+    public function finish(Request $request){
         //清process和ending資料
         $user_id = auth()->user()->id;
-        $game_delete = game_process::where('user_id',$user_id)->get();
-        $game_delete->delete();
-        $end_delete = game_ending::where('user_id',$user_id)->get();
-        $end_delete->delete();
+        $game_delete = game_process::where('user_id',$user_id)->delete();
+        $end_delete = game_ending::where('user_id',$user_id)->delete();
         //準備ending
         $intelligence = $request->intelligence;
         $wealth = $request->wealth;
